fix a changes bug, add run_sql on dataframe, add tests
