+++ tests/test_sql.py
@@ -8,21 +8,28 @@ import pytest
 from sql import (
     add_pk_to_create_table_string,
     add_pk_to_sqlite_table,
+    apply_changes,
     get_create_table_string,
     insert_rows,
+    run_sql,
 )
 
 
 @pytest.fixture()
-def conn():
-    conn_ = sqlite3.connect(":memory:")
-    pd.DataFrame(
+def df():
+    yield pd.DataFrame(
         [
             {"date": datetime(2020, 1, 1, 1, 1, 1), "desc": "abc", "amount": 123},
             {"date": datetime(2020, 1, 1, 1, 1, 2), "desc": "def", "amount": 456},
         ]
-    ).to_sql("test", conn_, index=False)
-    yield conn_
+    )
+
+
+@pytest.fixture()
+def conn(df):
+    with sqlite3.connect(":memory:") as conn_:
+        df.to_sql("test", conn_, index=False)
+        yield conn_
 
 
 CREATE_TABLE_STRING = """
@@ -95,4 +102,38 @@ def test_insert_rows(conn):
         {"date": "2020-01-01 01:01:03", "desc": "ggg", "amount": 777},
         {"date": "2020-01-01 01:01:04", "desc": "hhh", "amount": 888},
     ]
+    assert actual == expected
+
+
+def test_apply_changes():
+    df_initial = pd.DataFrame(
+        [
+            {"a": 123, "b": "abc", "pk": 1},
+            {"a": 456, "b": "def", "pk": 2},
+            {"a": 789, "b": "ghi", "pk": 3},
+        ]
+    )
+    df_changes = pd.DataFrame(
+        [
+            {"a": 999, "b": "zzz", "pk": 2},
+            {"a": 444, "b": "fff", "pk": 4},
+        ]
+    )
+    actual_df = apply_changes(df_initial, df_changes)
+    actual = sorted(actual_df.to_dict(orient="records"), key=lambda x: x["pk"])
+    expected = [
+        {"a": 123, "b": "abc", "pk": 1},
+        {"a": 999, "b": "zzz", "pk": 2},
+        {"a": 789, "b": "ghi", "pk": 3},
+        {"a": 444, "b": "fff", "pk": 4},
+    ]
+    assert actual == expected
+
+
+def test_run_sql(df):
+    actual_df = run_sql(
+        "SELECT desc,amount FROM expenses WHERE amount < 200", df, table_name="expenses"
+    )
+    actual = actual_df.to_dict(orient="records")
+    expected = [{"desc": "abc", "amount": 123}]
     assert actual == expected
